update 0.1.8
-Bezig met Authenticatie security
