--- src/main/java/com/project/lingo/Presentation/HtmlController.java
@@ -1,14 +1,0 @@
-package com.project.lingo.Presentation;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-public class HtmlController {
-    /*@RequestMapping("/")
-    public String welcome() {
-        return "login";
-    }*/
-
-
-}
