+++ src/main/java/com/project/lingo/securingweb/WebSecurityConfig.java
@@ -3,6 +3,7 @@ package com.project.lingo.securingweb;
 import com.project.lingo.Data.dao.UserDetailsServiceDao;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.http.HttpMethod;
 import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
 import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
@@ -38,11 +39,14 @@ public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
         auth.authenticationProvider(authenticationProvider());
     }
 
+    //NIET AF! Post doet het niet!
     @Override
     protected void configure(HttpSecurity http) throws Exception {
         http
                 .authorizeRequests()
                 .antMatchers("/", "/home", "/register").permitAll()
+                .antMatchers(HttpMethod.GET, "/api/**", "/api/**/**").permitAll()
+                .antMatchers(HttpMethod.POST, "/api/**/**").permitAll()
                 .antMatchers("/resources/**","/static/**", "/css/**", "/js/**", "/img/**", "/icon/**").permitAll()
                 .anyRequest().authenticated()
                 .and()
